+++ tests/Highcharts/ColorsTest.php
@@ -21,7 +21,7 @@ final class ColorsTest extends AbstractHighchartTestCase
     public function testColors(): void
     {
         $this->chart->colors = ['#FF0000', '#00FF00', '#0000FF'];
-        $regex = '/colors: \["#FF0000","#00FF00","#0000FF"\]/';
+        $regex = 'colors: ["#FF0000","#00FF00","#0000FF"]';
         $this->assertChartMatchesRegularExpression($regex);
     }
 }