+++ tests/Highcharts/ExportingTest.php
@@ -25,20 +25,20 @@ final class ExportingTest extends AbstractHighchartTestCase
     {
         // align option (string - left/center/right)
         $this->chart->exporting['buttons'] = ['exportButton' => ['align' => 'center']];
-        $regex = '/exporting: \{"buttons":\{"exportButton":\{"align":"center"\}\}\}/';
+        $regex = 'exporting: {"buttons":{"exportButton":{"align":"center"}}}';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->chart->exporting['buttons'] = ['printButton' => ['align' => 'center']];
-        $regex = '/exporting: \{"buttons":\{"printButton":\{"align":"center"\}\}\}/';
+        $regex = 'exporting: {"buttons":{"printButton":{"align":"center"}}}';
         $this->assertChartMatchesRegularExpression($regex);
 
         // backgroundColor option
         $this->chart->exporting['buttons'] = ['exportButton' => ['backgroundColor' => 'blue']];
-        $regex = '/exporting: \{"buttons":\{"exportButton":\{"backgroundColor":"blue"\}\}\}/';
+        $regex = 'exporting: {"buttons":{"exportButton":{"backgroundColor":"blue"}}}';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->chart->exporting['buttons'] = ['printButton' => ['backgroundColor' => 'blue']];
-        $regex = '/exporting: \{"buttons":\{"printButton":\{"backgroundColor":"blue"\}\}\}/';
+        $regex = 'exporting: {"buttons":{"printButton":{"backgroundColor":"blue"}}}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -48,11 +48,11 @@ final class ExportingTest extends AbstractHighchartTestCase
     public function testEnabled(): void
     {
         $this->chart->exporting['enabled'] = true;
-        $regex = '/exporting: \{"enabled":true\}/';
+        $regex = 'exporting: {"enabled":true}';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->chart->exporting['enabled'] = false;
-        $regex = '/exporting: \{"enabled":false\}/';
+        $regex = 'exporting: {"enabled":false}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -62,7 +62,7 @@ final class ExportingTest extends AbstractHighchartTestCase
     public function testFilename(): void
     {
         $this->chart->exporting['filename'] = 'graph';
-        $regex = '/exporting: \{"filename":"graph"\}/';
+        $regex = 'exporting: {"filename":"graph"}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -72,19 +72,19 @@ final class ExportingTest extends AbstractHighchartTestCase
     public function testType(): void
     {
         $this->chart->exporting['type'] = 'image/png';
-        $regex = '/exporting: \{"type":"image\/png"\}/';
+        $regex = 'exporting: {"type":"image/png"}';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->chart->exporting['type'] = 'image/jpeg';
-        $regex = '/exporting: \{"type":"image\/jpeg"\}/';
+        $regex = 'exporting: {"type":"image/jpeg"}';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->chart->exporting['type'] = 'application/pdf';
-        $regex = '/exporting: \{"type":"application\/pdf"\}/';
+        $regex = 'exporting: {"type":"application/pdf"}';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->chart->exporting['type'] = 'image/svg+xml';
-        $regex = '/exporting: \{"type":"image\/svg\+xml"\}/';
+        $regex = 'exporting: {"type":"image/svg+xml"}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -94,7 +94,7 @@ final class ExportingTest extends AbstractHighchartTestCase
     public function testUrl(): void
     {
         $this->chart->exporting['url'] = 'https://www.google.com';
-        $regex = '/exporting: \{"url":"https:\/\/www.google.com"\}/';
+        $regex = 'exporting: {"url":"https://www.google.com"}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -104,7 +104,7 @@ final class ExportingTest extends AbstractHighchartTestCase
     public function testWidth(): void
     {
         $this->chart->exporting['width'] = 300;
-        $regex = '/exporting: \{"width":300\}/';
+        $regex = 'exporting: {"width":300}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 }