+++ tests/Highcharts/LangTest.php
@@ -24,7 +24,7 @@ final class LangTest extends AbstractHighchartTestCase
     public function testMonths(): void
     {
         $this->chart->lang['months'] = ['Janvier', 'Fevrier', 'Mars', 'Avril', 'Mai', 'Juin',  'Juillet', 'Aout', 'Septembre', 'Octobre', 'Novembre', 'Decembre'];
-        $regex = '/lang: \{"months":\["Janvier","Fevrier","Mars","Avril","Mai","Juin","Juillet","Aout","Septembre","Octobre","Novembre","Decembre"\]\}/';
+        $regex = 'lang: {"months":["Janvier","Fevrier","Mars","Avril","Mai","Juin","Juillet","Aout","Septembre","Octobre","Novembre","Decembre"]}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -34,7 +34,7 @@ final class LangTest extends AbstractHighchartTestCase
     public function testShortMonths(): void
     {
         $this->chart->lang['shortMonths'] = ['Jan', 'Fev', 'Mars', 'Avril', 'Mai', 'Juin',  'Juil', 'Aout', 'Sept', 'Oct', 'Nov', 'Dec'];
-        $regex = '/lang: \{"shortMonths":\["Jan","Fev","Mars","Avril","Mai","Juin","Juil","Aout","Sept","Oct","Nov","Dec"\]\}/';
+        $regex = 'lang: {"shortMonths":["Jan","Fev","Mars","Avril","Mai","Juin","Juil","Aout","Sept","Oct","Nov","Dec"]}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -44,7 +44,7 @@ final class LangTest extends AbstractHighchartTestCase
     public function testWeekDays(): void
     {
         $this->chart->lang['weekdays'] = ['Dimanche', 'Lundi', 'Mardi', 'Mercredi', 'Jeudi', 'Vendredi', 'Samedi'];
-        $regex = '/lang: \{"weekdays":\["Dimanche","Lundi","Mardi","Mercredi","Jeudi","Vendredi","Samedi"\]\}/';
+        $regex = 'lang: {"weekdays":["Dimanche","Lundi","Mardi","Mercredi","Jeudi","Vendredi","Samedi"]}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 }