+++ tests/Highcharts/RenderToTest.php
@@ -20,14 +20,14 @@ final class RenderToTest extends AbstractHighchartTestCase
 {
     public function testWithDefault(): void
     {
-        $regex = '/const chart = new Highcharts/';
+        $regex = 'const chart = new Highcharts';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
     public function testWithValue(): void
     {
         $this->chart->chart['renderTo'] = 'myChart';
-        $regex = '/const myChart = new Highcharts/';
+        $regex = 'const myChart = new Highcharts';
         $this->assertChartMatchesRegularExpression($regex);
     }
 }