+++ tests/Highcharts/SeriesTest.php
@@ -30,10 +30,10 @@ final class SeriesTest extends AbstractHighchartTestCase
 
         $this->chart->series->merge($series);
 
-        $regex = '/\{"name":"Data #1","data":\[1,2,4,5,6,3,8\]\}/';
+        $regex = '{"name":"Data #1","data":[1,2,4,5,6,3,8]}';
         $this->assertChartMatchesRegularExpression($regex);
 
-        $regex = '/\{"name":"Data #2","data":\[7,3,5,1,6,5,9\]\}/';
+        $regex = '{"name":"Data #2","data":[7,3,5,1,6,5,9]}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 }