+++ tests/Highstock/CreditsTest.php
@@ -32,12 +32,12 @@ final class CreditsTest extends AbstractHighstockTestCase
         self::assertNotNull($this->credits);
         $this->credits['enabled'] = true;
         self::assertTrue($this->credits['enabled']);
-        $regex = '/"enabled":true/';
+        $regex = '"enabled":true';
         $this->assertChartMatchesRegularExpression($regex);
 
         $this->credits['enabled'] = false;
         self::assertFalse($this->credits['enabled']);
-        $regex = '/"enabled":false/';
+        $regex = '"enabled":false';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -62,7 +62,7 @@ final class CreditsTest extends AbstractHighstockTestCase
         self::assertNotNull($this->credits);
         $this->credits->position($position);
         self::assertSame($this->credits->position, $position);
-        $regex = '/"position":{"align":"right","x":-10,"verticalAlign":"bottom","y":-5}/';
+        $regex = '"position":{"align":"right","x":-10,"verticalAlign":"bottom","y":-5}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -77,7 +77,7 @@ final class CreditsTest extends AbstractHighstockTestCase
         self::assertNotNull($this->credits);
         $this->credits->style($style);
         self::assertSame($style, $this->credits->style);
-        $regex = '/"style":{"cursor":"pointer","color":"#909090","fontSize":"10px"}/';
+        $regex = '"style":{"cursor":"pointer","color":"#909090","fontSize":"10px"}';
         $this->assertChartMatchesRegularExpression($regex);
     }
 
@@ -88,7 +88,7 @@ final class CreditsTest extends AbstractHighstockTestCase
         $text = 'Highcharts.com';
         $this->credits->text($text);
         self::assertSame($text, $this->credits->text);
-        $regex = '/"text":"Highcharts.com"/';
+        $regex = '"text":"Highcharts.com"';
         $this->assertChartMatchesRegularExpression($regex);
     }
 }